move `getInvestmentOffer()` to corp object
